+++ decidim-verifications/spec/system/authorizations_spec.rb
@@ -135,6 +135,46 @@ describe "Authorizations", type: :system, with_authorization_workflows: ["dummy_
         end
       end
 
+      context "when the authorization is renewable" do
+        describe "and still not over the waiting period" do
+          let!(:authorization) do
+            create(:authorization, name: "dummy_authorization_handler", user: user, granted_at: 1.minute.ago)
+          end
+
+          it "can't be renewed yet" do
+            within_user_menu do
+              click_link "My account"
+            end
+
+            click_link "Authorizations"
+
+            within ".authorizations-list" do
+              expect(page).to have_no_link("Example authorization")
+              expect(page).to have_no_css(".authorization-renewable")
+            end
+          end
+        end
+
+        describe "and passed the time between renewals" do
+          let!(:authorization) do
+            create(:authorization, name: "dummy_authorization_handler", user: user, granted_at: 6.minutes.ago)
+          end
+
+          it "can be renewed" do
+            within_user_menu do
+              click_link "My account"
+            end
+
+            click_link "Authorizations"
+
+            within ".authorizations-list" do
+              expect(page).to have_link("Example authorization")
+              expect(page).to have_css(".authorization-renewable")
+            end
+          end
+        end
+      end
+
       context "when the authorization has not expired yet" do
         let!(:authorization) do
           create(:authorization, name: "dummy_authorization_handler", user: user, granted_at: 2.seconds.ago)